fix: store uploads in writable tmp dir, serve via PHP proxy

+++ modules/student/profile.php
@@ -20,9 +20,9 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['update_photo']) && $s
             $filename = 'student_' . $student['id'] . '_' . time() . '.' . $ext;
             $dest = ensure_upload_dir('students') . '/' . $filename;
             if (move_uploaded_file($_FILES['profile_image']['tmp_name'], $dest)) {
-                $old_path = ensure_upload_dir('students') . '/' . $student['profile_image'];
-                if ($student['profile_image'] && file_exists($old_path)) {
-                    unlink($old_path);
+                if ($student['profile_image']) {
+                    $old_path = get_upload_base() . '/students/' . $student['profile_image'];
+                    if (file_exists($old_path)) unlink($old_path);
                 }
                 db_query("UPDATE students SET profile_image=? WHERE id=?", [$filename, $student['id']]);
                 $student['profile_image'] = $filename;
